pyproject.toml
[project]
name = "crewai-mcp-newsletter"
version = "0.1.0"
description = "CrewAI multi-agent newsletter system with MCP email tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "crewai>=0.193.0",
  "fastmcp>=2.12,<3.0",
  "langchain-openai>=0.1.17",
  "openai>=1.0.0",
  "requests>=2.31.0",
  "python-dotenv>=1.0",
]

[tool.uv]
package = false